Look at actual refs to track if repos were updated
We are currently executing the update-hook regardless of whether there
were actual updates in the repositories we track via grok-dumb-pull.
This creates unnecessary churn, so look at ref-parse before and after
the git pull to figure out if there were any actual updates in the
repository before calling the update-hook.