Initial repository "fingerprint" implementation
A continuous problem is recognizing whether master and mirrors are the
same. We are relying on timestamps, but they are external to the git
repository itself and therefore aren't as reliable as data we can get
directly out of git. Comparing the output of "git show-ref" is one way
of comparing the state of two git repositories -- if the output is the
same, then the repositories should be identical, at least in terms of
refs.

Since the order of "git show-ref" is deterministic, it should be
reliable to use the hash of the output as a makeshift "repository
fingerprint", which we now rely on to make sure our mirrors match the
master.