Initial work to support multiple alternates
Also initial support for alternates modification. Needs more work and
testing, but appears to be doing the right thing in tests I've run. This
will additionally allow us to be less blunt when deleting repositories
-- we now should be able to recognize when we're deleting a repository
that is used as alternates for another repo and repack on the fly
accordingly (not coded yet, but at least feasible).