Reduce IO churn when comparing fingerprints
Rely on grokmirror.timestamp to do initial "check if this repo
has already been updated underneath us by another process"
as there can be cases when we unnecessarily fingerprint and
git update repositories when the old process operates on a
stale manifest.

Only really relevant when mirroring thousands of repositories
that are updated by an automated process several times an
hour.