Rewrite helper routines a Repository class
This allows me to take a clean break from python 2.6 and
use the cool new stuff from 2.7 and 3.x without having to
worry about sticking to 15-year-old python paradigms.

Added test cases for all of the helper methods, too.

This is a work very much in progress and everything is
pretty much broken. :)