Don't add failed clone repo to our manifest
If we tried to clone a repo, but actual git clone command failed, then
we should not add it to our manifest and let the next process attempt to
re-clone it. This also fixes a race condition that happens when a
non-grokmirror process creates an empty repo during grokmirror run.