Add safe prune support to grok-fsck
We've been paranoid about running gc in order to avoid potential repo
corruption due to the use of alternates between repositories. This
solution tries to be safe when running prune by first checking if any
other repositories we keep track of are using the repository being
pruned as their objects/info/alternates. If no, then it is safe to
prune, as old objects are not referenced anywhere else. If yes, then we
never prune for the fear of potentially causing badness.

This code also removes quick_checks functionality, as it has become less
useful with repacks and prunes. Small repos will get checked as part of
the regular schedule.

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>