Fix for potentiall mirror divergence
We shouldn't be relying on timestamps if we have fingerprint data,
otherwise there may be cases where mirrors diverge due to latest
timestamp being in place. This normally only happens if you pull the
manifest from one location, but actual repo from another location, which
may not actually have the latest git repos.

I only came across it when I was pulling manifest from master, but repos
themselves from another mirror that was a few minutes behind the master,
but this should be fixed anyway.

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>