Rewrite threaded processing again
I wasn't happy with using pools and async processing, so switch back to
simpler queues. Lots of changes in this commit, but importantly, this
implements listening on a socket to receive push updates.

Lots of documentation and testing left to do, but the end of the 2.0
rewrite is looming closer.

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>