Tweak how we match siblings to avoid snowballing
One of the problems I discovered with CAF repos, is that they like to
lump unrelated repos into a single "joint" repo. Once that happens, our
algorithm ended up merging objstore repos together -- which made sense
in terms of space savings, but this resulted in a snowballing effect
where standalone small repos ended up being stored with those huge
unrelated ones, impacting small repo performance.

I'm not sure this set of fixes is sufficient to prevent it, but the test
is currently running and I'll know more once it completes.

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>