Add merge strategy based on shared blob size
In rare occasions, there can be two repos that have different root
commits, but that share a significant number of their blobs (e.g. an
existing repo got squashed and re-imported). Try a strategy recognizing
such repos when 75% of their total blob size is shared.

Not sure this is sane, but worth trying out. For now, this will remain
and undocumented experimental feature.

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>