Add support to ignore refs for grok-manifest
With gerrit-3.x, the default behaviour is to hide refs/meta/* from
anonymous access, so when we calculate fingerprints directly in the
repository, they will be different on the replicas due to missing an
entire refname.

It's now possible to pass -r refs/meta/* (plus any number of others) to
create a "view" that matches what anonymous clients get.

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>